Add dynamic script compilation, construction calculations, command and reports upload, "current army" estimations for a given village, authentication via auth key UUID, TW map script for injecting vault data on hover, enable CORS on all endpoints, use Serilog for logging with rolling file store, add tracking of transactions per DB upload

+++ TW.Vault/Scaffold_Model/Ally.cs
@@ -5,13 +5,13 @@ namespace TW.Vault.Scaffold_Model
 {
     public partial class Ally
     {
-        public int TribeId { get; set; }
+        public long TribeId { get; set; }
         public string TribeName { get; set; }
         public string Tag { get; set; }
         public int? Members { get; set; }
         public int? Villages { get; set; }
         public long? Points { get; set; }
         public long? AllPoints { get; set; }
-        public int? TribeRank { get; set; }
+        public long? TribeRank { get; set; }
     }
 }
